+++ app/src/main/java/br/com/alura/aluvery/ui/screens/HomeScreen.kt
@@ -1,0 +1,36 @@
+package br.com.alura.aluvery.ui.screens
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import br.com.alura.aluvery.ui.components.ProductsSection
+
+@Composable
+fun HomeScreen() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .verticalScroll(rememberScrollState()),
+        verticalArrangement = Arrangement.spacedBy(16.dp, Alignment.Top)
+    ) {
+        Spacer(Modifier)
+        ProductsSection()
+        ProductsSection()
+        ProductsSection()
+        Spacer(Modifier)
+    }
+}
+
+@Preview(showSystemUi = true)
+@Composable
+private fun HomeScreenPreview() {
+    HomeScreen()
+}
